Add extra Python requirements
Requirements were missing on pyyaml, rospkg, and catkin_pkg, which
prevented this from installing and running cleanly in a virtual environment.

This also makes some modifications to the Travis CI build, which was failing
because some libraries (Tornado) require a newer version of the SSL module
than is included in the default version of Python (2.7.5).